Example GET username in frontend
