--- containers/vuejs/src/components/HtppRequests.vue
@@ -1,19 +1,0 @@
-<script setup lang="ts">
-import axios from 'axios'
-import { useRouter } from 'vue-router'
-
-const router = useRouter()
-
-async function get(path: string) {
-  const jwt = localStorage.getItem('jwt')
-  return await axios
-    .get(`http://localhost:4242/api/${path}`, { headers: { Authorization: `Bearer ${jwt}` } })
-    .then((response) => {
-      return response.data
-    })
-    .catch(() => {
-      localStorage.removeItem('jwt')
-      router.replace({ path: '/login' })
-    })
-}
-</script>
