+++ containers/vuejs/src/getSocket.ts
@@ -1,13 +1,13 @@
 import { io, Socket } from 'socket.io-client'
 
-let gameSocket: Socket
-let chatSocket: Socket
+export let gameSocket: Socket
+export let chatSocket: Socket
 
-const initGameSocket = (opts: any) => {
+export const initGameSocket = (opts: any) => {
   gameSocket = getSocket('/game', opts)
 }
 
-const initChatSocket = (opts: any) => {
+export const initChatSocket = (opts: any) => {
   chatSocket = getSocket('/chat', opts)
 }
 
@@ -15,5 +15,3 @@ const getSocket = (namespace: string, opts: any) => {
   const url = import.meta.env.VITE_ADDRESS + ':' + import.meta.env.VITE_BACKEND_PORT
   return io(url + namespace, opts)
 }
-
-export { gameSocket, chatSocket, initGameSocket, initChatSocket }
